requirements.txt
python-pptx
langchain
openai
python-dotenv
mcp